+++ src/components/OfficerSection/OfficerSection.scss
@@ -1,0 +1,42 @@
+@import '../../variables.scss';
+
+@keyframes gridSlideUpFadeIn {
+    from {
+        opacity: 0;
+        transform: translateY(20px);
+    }
+    to {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
+.officer-section {
+
+    .title {
+        font-family: $font-heading;
+        font-size: 5rem;
+        color: $White;
+        font-weight: 600;
+        text-align: center;
+        padding: 3rem 0;
+
+        @media screen and (max-width: 768px) {
+            font-size: 3.2rem;
+        }
+    }
+
+    .officer-grid {
+        display: flex;
+        flex-wrap: wrap;
+        gap: 3rem;
+        justify-content: center;
+        padding-top: 3rem;
+        padding-bottom: 3rem;
+        flex-basis: 3rem;
+        animation: gridSlideUpFadeIn 0.5s ease-out;
+        max-height: 2000px;
+        overflow: hidden;
+        transition: max-height 0.5s ease-out;
+    }
+}
